clean up

+++ deepnet/utils/network/iohandler.py
@@ -48,7 +48,7 @@ class IOHandler(object):
             kernel32 = ctypes.windll.kernel32
             kernel32.SetConsoleMode(kernel32.GetStdHandle(-11), 7)
 
-    def print_metrics(self, metrics: tuple or list, snapshot_step: int):
+    def print_metrics(self, metrics: tuple or list, snapshot_step: float):
         if snapshot_step < time.time() - self._metric_last_update:
             self._print_metrics(metrics=metrics)
             self._metric_last_update = time.time()
